host/pyproject.toml
[project]
name = "applebridge"
version = "0.1.0"
description = "Connect Claude to classic Mac MPW shell via emulator"
authors = [{name = "AppleBridge Contributors"}]
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.39.0",
    "pillow>=10.0.0",
]

[project.scripts]
applebridge = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
